Add files via upload
I cant believe this is my first offsite save of all this stuff i have worked on off and on since 2011. Much of this is from elsewhere (siphash!)
I'm trying to work on a secured remote to host connection, this is a first pass, but misses the part about validating the host, and Im a long ways from using a single credential for a single entity.